Improved performance of some PluginData objects Fixed PluginConfigSectionHandler not creating new Enabled section properly for Factions & Towny

+++ PlanPluginBridge/src/main/java/com/djrapitops/pluginbridge/plan/advancedachievements/AdvancedAchievementsData.java
@@ -60,7 +60,7 @@ public class AdvancedAchievementsData extends PluginData {
     }
 
     private long getTotal(Map<UUID, Integer> totalAchievements) {
-        return MathUtils.sumLong(totalAchievements.values().stream().map(i -> i));
+        return MathUtils.sumLong(totalAchievements.values().stream().map(i -> (long) i));
     }
 
     private void refreshTotalAchievements() {
